fix: circular import and BenchmarkRunner
